minor refactor to deleteUser to (superadmin can't delete themselves)

+++ src/resolvers/Mutation/deleteUser.js
@@ -1,5 +1,6 @@
 import getUserId from '../../utils/getUserId'
 import isAdmin from '../../utils/isAdmin'
+import isSuperadmin from '../../utils/isSuperadmin'
 
 const deleteUser = async (parent, { id }, { prisma, request }, info) => {
   // Get user ID from token (will throw an error if unauthenticated)
@@ -8,6 +9,11 @@ const deleteUser = async (parent, { id }, { prisma, request }, info) => {
 
   // If id is not given, delete connected user
   if (!id) {
+    // Superadmin cannot delete themself!
+    if (isSuperadmin(connectedUser)) {
+      throw new Error('Unable to delete superadmin')
+    }
+
     return prisma.mutation.deleteUser(
       {
         where: {
@@ -18,13 +24,24 @@ const deleteUser = async (parent, { id }, { prisma, request }, info) => {
     )
   }
 
+  const userToDelete = await prisma.query.user({ where: { id } })
+
+  // Nobody can delete superadmin
+  if (isSuperadmin(userToDelete)) {
+    throw new Error('Unable to delete superadmin')
+  }
+
   // If id is given and connected user is admin, delete given user
   if (isAdmin(connectedUser)) {
+    // Admin cannot delete other admin
+    if (!isSuperadmin(connectedUser) && isAdmin(userToDelete)) {
+      throw new Error('Unable to delete admin')
+    }
     return prisma.mutation.deleteUser({ where: { id } }, info)
   }
 
   // If id is given and connected user is not admin, throw an error
-  throw new Error('You cannot delete a user')
+  throw new Error('Unable to delete user')
 }
 
 export { deleteUser as default }
